requirements.txt
# Core ML / NLP
sentence-transformers==2.7.0
faiss-cpu==1.8.0
torch==2.2.2
transformers==4.40.0

# Google Gemini
google-generativeai==0.5.4

# LangChain
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.3

# MongoDB
pymongo==4.7.2
motor==3.4.0

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Frontend
streamlit==1.35.0

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.4
loguru==0.7.2
tiktoken==0.7.0
weaviate-client==4.6.5